+++ lib/sequel-rails.rb
@@ -1,6 +1,1 @@
-module Rails
-  module Sequel
-  end
-end
-
-require 'sequel-rails/railtie'
+require "sequel_rails"