+++ lib/sequel-rails/storage/mysql2.rb
@@ -1,8 +1,0 @@
-module Rails
-  module Sequel
-    module Storage
-      class Mysql2 < Mysql
-      end
-    end
-  end
-end