+++ lib/sequel_rails.rb
@@ -1,0 +1,5 @@
+module SequelRails
+  VERSION = "0.4.0.pre"
+end
+
+require 'sequel_rails/railtie'