+++ lib/sequel_rails/railties/i18n_support.rb
@@ -1,0 +1,10 @@
+module SequelRails
+
+  module I18nSupport
+    # Set the i18n scope to overwrite ActiveModel.
+    def i18n_scope #:nodoc:
+      :sequel
+    end
+  end
+
+end