+++ lib/sequel_rails/storage/mysql2.rb
@@ -1,0 +1,6 @@
+module SequelRails
+  module Storage
+    class Mysql2 < Mysql
+    end
+  end
+end